bluesky/requirements.txt
beautifulsoup4==4.12.2
Jinja2==3.1.2
linkpreview==0.7.0
lxml==4.9.3
python-liquid==1.10.0
requests==2.31.0
urllib3==1.26.16

